Cleanup, ready for initial review
