+++ packages/devtools_app/lib/src/debugger/program_explorer_controller.dart
@@ -1,121 +1,68 @@
+// Copyright 2021 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
 import 'dart:async';
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:vm_service/vm_service.dart';
 
+import '../auto_dispose.dart';
 import '../globals.dart';
 import '../trees.dart';
 import '../utils.dart';
-import '../version.dart';
-import 'object_tree_selector.dart';
+import 'program_explorer_model.dart';
 
-class VMServiceLibraryContents {
-  const VMServiceLibraryContents({
-    this.lib,
-    this.classes,
-    this.functions,
-    this.fields,
-  });
-
-  final Library lib;
-  final List<ClassRef> classes;
-  final List<FuncRef> functions;
-  final List<FieldRef> fields;
-
-  static Future<VMServiceLibraryContents> getLibraryContents(
-      LibraryRef libRef) async {
-    final isolateId = serviceManager.isolateManager.selectedIsolate.value.id;
-    final service = serviceManager.service;
-
-    final lib = await service.getObject(isolateId, libRef.id) as Library;
-    var classes = <ClassRef>[];
-    var functions = <FuncRef>[];
-    var fields = <FieldRef>[];
-
-    if (!displayLibraryExplorer) {
-      classes.addAll(lib.classes);
-      functions.addAll(lib.functions);
-      fields.addAll(lib.variables);
-
-      // Before 3.46, ClassRef, FuncRef, and FieldRef didn't contain location
-      // information and couldn't be mapped to their parent scripts. For older
-      // versions of the protocol, we need to request the full objects for
-      // everything. We'll avoid doing this for versions >= 3.46 and lazily
-      // populate the tree with full instances as the user navigates.
-      if (!await service.isProtocolVersionSupported(
-        supportedVersion: SemanticVersion(major: 3, minor: 46),
-      )) {
-        final classesRequests = lib.classes.map(
-          (clazz) async =>
-              await service.getObject(isolateId, clazz.id) as Class,
-        );
-
-        classes = await Future.wait(classesRequests);
-      }
-
-      final funcsRequests = lib.functions.map(
-        (func) async => await service.getObject(isolateId, func.id) as Func,
-      );
-      functions = await Future.wait(funcsRequests);
-
-      final fieldsRequests = lib.variables.map(
-        (field) async => await service.getObject(isolateId, field.id) as Field,
-      );
-      fields = await Future.wait(fieldsRequests);
-    }
-
-    // Remove scripts pulled into libraries via mixins.
-    lib.scripts.removeWhere((e) => !e.uri.contains(lib.uri));
-
-    return VMServiceLibraryContents(
-      lib: lib,
-      classes: classes,
-      functions: functions,
-      fields: fields,
+class ProgramExplorerController extends DisposableController
+    with AutoDisposeControllerMixin {
+  ProgramExplorerController() {
+    addAutoDisposeListener(
+      serviceManager.isolateManager.selectedIsolate,
+      () => refresh(),
     );
   }
-}
 
-class ObjectTreeController {
-  final programStructure = <VMServiceLibraryContents>[];
+  /// A list of objects containing the contents of each library.
+  final _programStructure = <VMServiceLibraryContents>[];
 
+  /// The total number of selectable objects in the unfiltered tree.
   ValueListenable<int> get objectCount => _objectCount;
   final _objectCount = ValueNotifier<int>(0);
 
+  /// The total number of selectable objects in the filtered tree.
   ValueListenable<int> get filteredObjectCount => _filteredObjectCount;
   final _filteredObjectCount = ValueNotifier<int>(0);
 
-  ValueListenable<VMServiceObjectNode> get selected => _selected;
-  final _selected = ValueNotifier<VMServiceObjectNode>(null);
+  /// The currently selected node.
+  VMServiceObjectNode _selected;
 
+  /// The processed roots of the tree.
   ValueListenable<List<VMServiceObjectNode>> get rootObjectNodes =>
       _rootObjectNodes;
   final _rootObjectNodes = ValueNotifier<List<VMServiceObjectNode>>([]);
 
-  // Cache of object IDs to their containing library to allow for easier
-  // refreshing of library content, particularly for scripts.
+  /// Cache of object IDs to their containing library to allow for easier
+  /// refreshing of library content, particularly for scripts.
   final _objectIdToLibrary = <String, LibraryRef>{};
 
   IsolateRef _isolate;
 
-  ValueListenable<bool> get initializationListenable =>
-      _initializationListenable;
+  /// Notifies that the controller has finished initializing.
+  ValueListenable<bool> get initialized => _initializationListenable;
   final _initializationListenable = ValueNotifier<bool>(false);
   bool _initializing = false;
 
+  /// Attaches filtering information to package:vm_service objects.
   final _shouldFilterExpando = Expando<bool>('shouldFilter');
+
+  /// The current list of library contents that match the current filter. This
+  /// is a strict subset of contents stored in [_programStructure].
   List<VMServiceLibraryContents> _filteredItems;
 
+  /// Returns true if [function] is a getter or setter that was not explicitly
+  /// defined (e.g., `int foo` creates `int get foo` and `set foo(int)`).
   static bool _isSyntheticAccessor(FuncRef function, List<FieldRef> fields) {
-    if (serviceManager.service.isProtocolVersionSupportedNow(
-      supportedVersion: SemanticVersion(major: 3, minor: 47),
-    )) {
-      // TODO(bkonyi)
-      /*if (function.synthetic) {
-        return true;
-      }*/
-    }
     for (final field in fields) {
       if (function.name == field.name || function.name == '${field.name}=') {
         return true;
@@ -125,34 +72,36 @@ class ObjectTreeController {
   }
 
   /// Initializes the program structure.
+  // TODO(bkonyi): reinitialize after hot reload.
   Future<void> initialize() async {
     if (_initializing) {
       return;
     }
     _initializing = true;
-    final libraries = await Future.wait(
-      serviceManager
-          .isolateManager.mainIsolateDebuggerState.isolateNow.libraries
-          .map(
-        (lib) => VMServiceLibraryContents.getLibraryContents(lib),
-      ),
-    );
     _isolate = serviceManager.isolateManager.selectedIsolate.value;
-
-    int count = 0;
+    final libraries = await Future.wait(
+      serviceManager.isolateManager
+          .isolateDebuggerState(_isolate)
+          .isolateNow
+          .libraries
+          .map(
+            (lib) => VMServiceLibraryContents.getLibraryContents(lib),
+          ),
+    );
 
     void mapIdsToLibrary(LibraryRef lib, Iterable<ObjRef> objs) {
       for (final e in objs) {
         _objectIdToLibrary[e.id] = lib;
-        count++;
       }
     }
 
     for (final libContents in libraries) {
       final lib = libContents.lib;
+      mapIdsToLibrary(lib, [lib]);
       mapIdsToLibrary(lib, lib.scripts);
       mapIdsToLibrary(lib, libContents.classes);
       mapIdsToLibrary(lib, libContents.fields);
+
       // Filter out synthetic getters/setters
       final filteredFunctions = libContents.functions.where(
         (e) => !_isSyntheticAccessor(e, libContents.fields),
@@ -160,39 +109,54 @@ class ObjectTreeController {
       libContents.functions.clear();
       libContents.functions.addAll(filteredFunctions);
       mapIdsToLibrary(lib, libContents.functions);
+
+      // Account for entries in library parts.
+      mapIdsToLibrary(lib, libContents.functions.map((e) => e.location.script));
+      mapIdsToLibrary(lib, libContents.classes.map((e) => e.location.script));
+      mapIdsToLibrary(lib, libContents.fields.map((e) => e.location.script));
     }
 
-    _objectCount.value += count;
-    programStructure.addAll(libraries.cast<VMServiceLibraryContents>());
+    _programStructure.addAll(libraries);
 
     // Build the initial tree.
     updateVisibleNodes();
+
+    // Initial tree contains all nodes, so we can use the filtered object count
+    // based on the actual tree structure instead of trying to count here using
+    // logic not based on the tree structure itself.
+    _objectCount.value = _filteredObjectCount.value;
     _initializationListenable.value = true;
   }
 
-/*
-  Future<void> refresh() {
+  /// Clears controller state and re-initializes.
+  Future<void> refresh() async {
     _objectIdToLibrary.clear();
-    _selected.value = null;
-    programStructure.clear();
-    return initialize();
+    _selected = null;
+    _initializationListenable.value = false;
+    _initializing = false;
+    _programStructure.clear();
+    _objectCount.value = 0;
+    _filteredObjectCount.value = 0;
+    return await initialize();
   }
-*/
+
+  /// Marks [node] as the currently selected node, clearing the selection state
+  /// of any currently selected node.
   void selectNode(VMServiceObjectNode node) {
     if (!node.isSelectable) {
       return;
     }
-    if (_selected.value != node) {
+    if (_selected != node) {
       node.isSelected = true;
-      _selected.value?.isSelected = false;
-      _selected.value = node;
+      _selected?.isSelected = false;
+      _selected = node;
     }
   }
 
   /// Rebuilds the tree nodes, only creating nodes for objects that match
   /// [filterText].
   void updateVisibleNodes([String filterText = '']) {
-    for (final ref in programStructure) {
+    for (final ref in _programStructure) {
       bool includeLib = false;
       if (ref.lib.uri.caseInsensitiveFuzzyMatch(filterText)) {
         includeLib = true;
@@ -230,8 +194,9 @@ class ObjectTreeController {
       }
       _shouldFilterExpando[ref.lib] = includeLib;
     }
-    _filteredItems =
-        programStructure.where((ref) => _shouldFilterExpando[ref.lib]).toList();
+    _filteredItems = _programStructure
+        .where((ref) => _shouldFilterExpando[ref.lib])
+        .toList();
 
     // Remove the cached value here; it'll be re-computed the next time we need
     // it.
@@ -240,19 +205,6 @@ class ObjectTreeController {
       _shouldFilterExpando,
     );
     _updateFilteredObjectCount();
-  }
-
-  /// Iterates over the tree and counts the number of selectable nodes.
-  void _updateFilteredObjectCount() {
-    _filteredObjectCount.value = rootObjectNodes.value.fold(0, (prev, e) {
-      int count = 0;
-      breadthFirstTraversal<VMServiceObjectNode>(e, action: (e) {
-        if (e.isSelectable) {
-          count++;
-        }
-      });
-      return prev + count;
-    });
   }
 
   /// Updates `node` with a fully populated VM service [Obj].
@@ -304,7 +256,7 @@ class ObjectTreeController {
     // the full object.
     updatedObj = await service.getObject(_isolate.id, updatedObj.id);
 
-    final library = programStructure.firstWhere(
+    final library = _programStructure.firstWhere(
       (e) => e.lib.uri == lib.uri,
     );
     if (updatedObj is Class) {
@@ -357,5 +309,18 @@ class ObjectTreeController {
     // We might have expanded a Class and added new entries to the tree. Make
     // sure we account for these new nodes to the filtered object count.
     _updateFilteredObjectCount();
+  }
+
+  /// Iterates over the tree and counts the number of selectable nodes.
+  void _updateFilteredObjectCount() {
+    _filteredObjectCount.value = rootObjectNodes.value.fold(0, (prev, e) {
+      int count = 0;
+      breadthFirstTraversal<VMServiceObjectNode>(e, action: (e) {
+        if (e.isSelectable) {
+          count++;
+        }
+      });
+      return prev + count;
+    });
   }
 }
